Renamed SwithcerV2 -> Switcher
