+++ src/components/Switcher/Switcher.js
@@ -1,72 +1,21 @@
 import React from "react"
-import { SwitcherItem } from "."
-import {
-  BsFillInboxFill,
-  BsCalendar3,
-  BsCalendarDate,
-  BsFillFlagFill,
-} from "react-icons/bs"
+import { Wrapper, Body, Header, Icon, Length } from "./Switcher.styled"
 
-const Switcher = () => {
-  const icon = 0,
-    color = 2,
-    length = 0
+const SwitcherV2 = props => {
+  const { icon, length, color, bgColor } = props
   return (
-    <div className="d-flex flex-column">
-      <div className="d-flex">
-        <SwitcherItem
-          color={"switcher-text"}
-          colorFocus={color}
-          bgColor={"bg-color-gray-lighten"}
-          bgColorFocus={"bg-color-gray"}
-          Icon={BsFillInboxFill}
-          iconColor={"color-white"}
-          iconBackgroundColor={"bg-color-gray"}
-          iconColorFocus={icon}
-          text={"All"}
-          length={length}
-        />
-        <SwitcherItem
-          color={"switcher-text"}
-          colorFocus={color}
-          bgColor={"bg-color-gray-lighten"}
-          bgColorFocus={"bg-color-primary"}
-          Icon={BsCalendar3}
-          iconColor={"color-white"}
-          iconBackgroundColor={"bg-color-danger"}
-          iconColorFocus={icon}
-          text={"Calendar"}
-          length={length}
-        />
-      </div>
-      <div className="d-flex">
-        <SwitcherItem
-          color={"switcher-text"}
-          colorFocus={color}
-          bgColor={"bg-color-gray-lighten"}
-          bgColorFocus={"bg-color-danger"}
-          Icon={BsCalendarDate}
-          iconColor={"color-white"}
-          iconBackgroundColor={"bg-color-primary"}
-          iconColorFocus={icon}
-          text={"Today"}
-          length={length}
-        />
-        <SwitcherItem
-          color={"switcher-text"}
-          colorFocus={color}
-          bgColor={"bg-color-gray-lighten"}
-          bgColorFocus={"bg-color-warning"}
-          Icon={BsFillFlagFill}
-          iconColor={"color-white"}
-          iconBackgroundColor={"bg-color-warning"}
-          iconColorFocus={icon}
-          text={"Flagged"}
-          length={length}
-        />
-      </div>
-    </div>
+    <>
+      <Wrapper>
+        <Body tabIndex="1" color={color} bgColor={bgColor}>
+          <Header>
+            <Icon color={color}>{icon}</Icon>
+            <Length>{length}</Length>
+          </Header>
+          {props.children}
+        </Body>
+      </Wrapper>
+    </>
   )
 }
 
-export default Switcher
+export default SwitcherV2
