+++ src/components/Switcher/index.js
@@ -1,3 +1,1 @@
 export { default as Switcher } from "./Switcher"
-export { default as SwitcherItem } from "./SwitcherItem"
-export { default as SwitcherV2 } from "./SwitcherV2"
